Fix integration test compile + S3 generateDownloadURL
ContextFileIT / DriveFileUploadIT compile failures came from the
fileSize schema switch to integer/int64 — the generated setter/getter
is now Integer. Replace the double literals with ints and the
assertEquals(double, ...) sites with intValue() so the (int, int)
overload resolves unambiguously.

Also override S3AssetService.generateDownloadURL to return a
short-lived presigned URL (mirroring AzureAssetService) instead of
inheriting the default, which would return the raw S3 key from
asset.url. Addresses review 4151282021.

+++ openmetadata-integration-tests/src/test/java/org/openmetadata/it/drive/ContextFileIT.java
@@ -57,7 +57,7 @@ class ContextFileIT {
             .withName(ns.prefix("report-pdf"))
             .withDisplayName("Annual Report 2023")
             .withFileType(ContextFileType.PDF)
-            .withFileSize(4200000.0)
+            .withFileSize(4200000)
             .withContentType("application/pdf")
             .withFileExtension("pdf")
             .withProcessingStatus(ProcessingStatus.Uploaded);
@@ -66,7 +66,7 @@ class ContextFileIT {
     assertNotNull(file.getId());
     assertEquals("Annual Report 2023", file.getDisplayName());
     assertEquals(ContextFileType.PDF, file.getFileType());
-    assertEquals(4200000.0, file.getFileSize());
+    assertEquals(4200000, file.getFileSize().intValue());
   }
 
   @Test
@@ -78,7 +78,7 @@ class ContextFileIT {
             .withName(ns.prefix("pricing-xlsx"))
             .withDisplayName("Product Pricing")
             .withFileType(ContextFileType.Spreadsheet)
-            .withFileSize(128000.0)
+            .withFileSize(128000)
             .withContentType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet")
             .withFileExtension("xlsx")
             .withProcessingStatus(ProcessingStatus.Uploaded);
